# frontend/compassClient/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/timestamp":{"put":{"tags":["timestamp-controller"],"operationId":"putTimestamp","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTimestampDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetTimestampDto"}}}}}},"post":{"tags":["timestamp-controller"],"operationId":"createTimestamp","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTimestampDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetTimestampDto"}}}}}}},"/api/daysheet":{"put":{"tags":["day-sheet-controller"],"operationId":"updateDay","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateDaySheetDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetDaySheetDto"}}}}}},"post":{"tags":["day-sheet-controller"],"operationId":"createDaySheet","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDaySheetDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetDaySheetDto"}}}}}}},"/api/timestamp/{id}":{"get":{"tags":["timestamp-controller"],"operationId":"getTimestampById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetTimestampDto"}}}}}},"delete":{"tags":["timestamp-controller"],"operationId":"deleteTimestamp","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/daysheet/{id}":{"get":{"tags":["day-sheet-controller"],"operationId":"getDaySheetById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetDaySheetDto"}}}}}}},"/api/daysheet/getByDate/{date}":{"get":{"tags":["day-sheet-controller"],"operationId":"getDaySheetById_1","parameters":[{"name":"date","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetDaySheetDto"}}}}}}},"/":{"get":{"tags":["base-controller"],"operationId":"index","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"UpdateTimestampDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"day_sheet_id":{"type":"integer","format":"int64"},"start_time":{"type":"string","format":"date-time"},"end_time":{"type":"string","format":"date-time"}}},"GetTimestampDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"day_sheet_id":{"type":"integer","format":"int64"},"start_time":{"type":"string","format":"date-time"},"end_time":{"type":"string","format":"date-time"}}},"UpdateDaySheetDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"date":{"type":"string","format":"date-time"},"day_report":{"type":"string"}}},"GetDaySheetDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"date":{"type":"string","format":"date-time"},"day_report":{"type":"string"},"confirmed":{"type":"boolean"},"timestamps":{"type":"array","items":{"$ref":"#/components/schemas/Timestamp"}}}},"Timestamp":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"startTime":{"type":"string","format":"date-time"},"endTime":{"type":"string","format":"date-time"}}},"CreateTimestampDto":{"type":"object","properties":{"day_sheet_id":{"type":"integer","format":"int64"},"start_time":{"type":"string","format":"date-time"},"end_time":{"type":"string","format":"date-time"}}},"CreateDaySheetDto":{"type":"object","properties":{"date":{"type":"string","format":"date-time"},"day_report":{"type":"string"}}}}}}